fixed pooling bug and added delta scroll

+++ Assets/UnitySoFunctional/Controls/PoolManager.cs
@@ -7,7 +7,7 @@ namespace DragonDogStudios.UnitySoFunctional.Controls
     {
         public event Action PooledItemsUpdated;
         public abstract float LogItemHeight { get; }
-        public abstract float Count { get; }
+        public abstract int Count { get; }
         public abstract bool SnapToBottom { get; set; }
 
         protected void InvokePooledItemsUpdated()
